+++ Node.h
@@ -25,7 +25,7 @@ struct TreeNode {
 #include "Variables.h"
 
 ErrorCode node_init     (TreeNode** node);
-ErrorCode node_init_num (TreeNode** node, int value);
+ErrorCode node_init_num (TreeNode** node, double value);
 ErrorCode node_delete   (TreeNode* node);
 ErrorCode node_copy     (TreeNode* node, TreeNode** new_node);
 ErrorCode node_insert_op(TreeNode** dest, TypeOperator operator_,
